Исправил, теперь все нормально апдейтится (правда с помощью костыля сделал) + добавил ковертер + подрубил messageBox'ы

+++ ForwardTestTask/ViewModels/NoteViewModel.cs
@@ -1,8 +1,11 @@
-﻿using ForwardTestTask.Core.Services.Interfaces;
+﻿using Avalonia.Controls;
+
+using ForwardTestTask.Core.Services.Interfaces;
 using ForwardTestTask.Domain.Entities;
-using ForwardTestTask.Presentation.Models;
+using ForwardTestTask.Presentation.MessageBoxes;
 
 using ReactiveUI;
+using ReactiveUI.Fody.Helpers;
 
 using System;
 using System.Collections.Generic;
@@ -15,34 +18,50 @@ public class NoteViewModel : ViewModelBase
 {
     private readonly INoteService _noteService;
 
-    public NoteViewModel(INoteService noteService)
+    public NoteViewModel(INoteService noteService, Window window)
     {
         _noteService = noteService;
-        Notes = _noteService.Notes;
-        Notes.Subscribe(x =>
-        {
-        });
-        AddNoteCommand = ReactiveCommand.CreateFromTask<AddNoteModel>(AddNote);
-        DeleteNoteCommand = ReactiveCommand.Create<Guid>(async (guid) => await DeleteNote(guid));
-        EditNoteCommand = ReactiveCommand.Create<EditNoteModel>(async (_) => await EditNote(_));
+        _noteService.Notes.Subscribe(UpdateCollection);
+
+        Window = window;
+
+        AddNoteCommand = ReactiveCommand.CreateFromTask(AddNote);
+        DeleteNoteCommand = ReactiveCommand.CreateFromTask<Guid>(DeleteNote);
+        EditNoteCommand = ReactiveCommand.CreateFromTask<EditNoteModel>(EditNote);
     }
 
+    #region Properties
+    [Reactive] public IEnumerable<Note> Notes { get; private set; }
+    public Window Window { get; }
+    #endregion
+
     #region Commands
-    public IObservable<IList<Note>> Notes { get; }
-    public ReactiveCommand<AddNoteModel, Unit> AddNoteCommand { get; }
+    public ReactiveCommand<Unit, Unit> AddNoteCommand { get; }
     public ReactiveCommand<Guid, Unit> DeleteNoteCommand { get; set; }
     public ReactiveCommand<EditNoteModel, Unit> EditNoteCommand { get; }
     #endregion
 
     #region Methods
-    private async Task AddNote(AddNoteModel addNoteModel) => await _noteService.AddAsync(addNoteModel);
+    private async Task AddNote() => await _noteService.AddAsync(null);
 
     private async Task DeleteNote(Guid guid)
     {
-        //Message
-        await _noteService.DeleteAsync(guid);
+        var result = await MessageBoxHelper.ShowConfirmMessageBoxAsync(Window, "Вы уверены, что хотите удалить заметку?");
+        if (result)
+            await _noteService.DeleteAsync(guid);
     }
 
     private async Task EditNote(EditNoteModel editNoteModel) => await _noteService.EditAsync(editNoteModel);
+
+    /// <summary>
+    ///Костыль, который нужен чтобы нормально обновлялась вью: если не меняется ссылка у отображаемого объект,
+    ///то ItemsControl (как и ListBox) не изменяются
+    /// </summary>
+    /// <param name="source"></param>
+    private void UpdateCollection(IEnumerable<Note> source)
+    {
+        Notes = null!;
+        Notes = source;
+    }
     #endregion
 }
